Component refactoring
