+++ src/pages/HomePage.tsx
@@ -5,25 +5,18 @@ import '../styles/Custom.css'
 import { addMedia } from '../actions'
 import Loader from '../components/Loader'
 import Sidebar from '../components/Sidebar'
-import Search from '../components/Search'
-import Swiper from 'react-id-swiper';
+import Tags from '../components/Tags'
 
 
 import url, { mediaTypes, elementsOnLoad } from '../constants'
 import { shouldLazyLoad, gaEvents } from '../common_function'
-import SearchMobile from '../components/SearchMobile'
-import 'swiper/css/swiper.css';
+import Header from '../common/Header'
 
 //----------------------------------------------------------
 
 //----------------------------------------------------------
 
 const HomePage = () => {
-
-    const params = {
-        slidesPerView: 7,
-        spaceBetween: 15,
-    }
 
     const { mediaById, media }: { mediaById: any; media: number[] } = useSelector(
         (state: any) => ({
@@ -114,167 +107,119 @@ const HomePage = () => {
 
     return (
         <>
-            {/* {mediaSorted.length === 0 && <Loader />} */}
-            {mediaSorted.length >= 0 && (
-                <div className="flex flex-col h-full lg:flex-row">
-                    <Sidebar />
-                    <div className="w-full pb-8 vod-content">
-                        <div className="flex items-center justify-between px-8 py-3 shadow-md sm:flex-row ">
-                            <div className="flex-initial md:pl-4">
-                                <svg width="101" height="29" viewBox="0 0 101 29" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <path d="M0 14.0931C0 21.8459 0.0191652 28.1862 0.0447188 28.1862C0.0830491 28.1862 5.28959 27.2211 5.92843 27.0933L6.26063 27.0293V21.8523V16.6816H8.24103H10.2214V15.7357C10.2214 15.218 10.2406 13.8694 10.2661 12.7317L10.3045 10.6737H8.28575H6.26063V8.40473V6.13577H8.59239C9.87646 6.13577 10.9242 6.12299 10.9242 6.10381C10.9242 6.03351 11.4352 0.57523 11.4672 0.300398L11.5055 0H5.74955H0V14.0931Z" fill="#666FE4" />
-                                    <path d="M95.5573 0.824446L90.4594 1.24628V13.978V26.7097L90.6702 26.7417C90.8554 26.7736 100.726 27.4895 100.924 27.4831C100.981 27.4831 101 26.8056 101 24.3193V21.1556H98.8281H96.6561V18.8547V16.5537H98.5087H100.361V13.8054V11.0571H98.4448H96.5283V9.23556V7.41401L96.6753 7.38844C96.7519 7.36927 97.4035 7.3437 98.1254 7.31814C98.8473 7.29896 99.7864 7.267 100.221 7.24144L101 7.20309V3.79646V0.383438L100.828 0.38983C100.726 0.396221 98.3554 0.587963 95.5573 0.824446Z" fill="#666FE4" />
-                                    <path d="M62.8303 1.70625C62.8942 2.34539 63.6225 11.9773 63.6097 11.9901C63.6033 12.0028 62.6642 9.69553 61.5207 6.86413C60.3836 4.03273 59.4253 1.68707 59.3997 1.64233C59.3614 1.57203 57.5982 1.89799 54.7043 2.51157L54.4296 2.56909V13.5559V24.5428H57.2724H60.1152L60.0833 24.0826C60.0194 23.2262 59.783 16.1445 59.8086 15.8633C59.8278 15.614 60.0194 16.0486 61.8337 20.4331C62.9325 23.0983 63.8652 25.3034 63.8908 25.3417C63.9291 25.3801 64.8171 25.3226 66.3248 25.1947C67.6344 25.0797 68.7076 24.9774 68.7076 24.971C68.714 24.9646 68.6821 20.3564 68.6374 14.732C68.5926 9.10752 68.5543 3.79624 68.5543 2.92062L68.5479 1.34194H65.6731H62.792L62.8303 1.70625Z" fill="#666FE4" />
-                                    <path d="M17.8577 1.73105C14.0374 1.98671 11.9356 4.6008 11.2712 9.90569C11.1371 11.005 11.0796 14.7376 11.1754 16.0415C11.4948 20.2726 12.6383 23.0209 14.6698 24.4206C15.7623 25.1684 17.4041 25.5711 19.0267 25.4752C20.873 25.3729 22.1315 24.8488 23.2622 23.7176C24.1885 22.7908 24.6932 21.9216 25.1787 20.4388C25.8559 18.3679 26.1626 15.2745 26.0092 12.0532C25.8368 8.44844 25.2937 6.20505 24.1758 4.48576C23.1089 2.84955 21.4224 1.87166 19.4547 1.75023C19.1417 1.73105 18.8095 1.71188 18.7201 1.6991C18.6306 1.69271 18.2473 1.70549 17.8577 1.73105ZM18.7584 6.75471C19.2248 7.25964 19.4228 9.27294 19.4228 13.5808C19.4228 18.8856 19.0906 21.0907 18.3623 20.6177C18.0493 20.4068 17.8129 19.3778 17.6979 17.6713C17.6149 16.4249 17.6149 10.7174 17.6979 9.5222C17.8193 7.83486 18.0365 6.90172 18.3623 6.67802C18.5412 6.55019 18.5732 6.55658 18.7584 6.75471Z" fill="#666FE4" />
-                                    <path d="M26.8309 13.3506V24.9893H29.7632H32.6891L32.7274 24.1776C32.753 23.7238 32.8041 22.2538 32.8424 20.8988C32.8807 19.5439 32.9318 18.355 32.9574 18.2592C32.9893 18.125 33.264 18.9878 34.1073 21.9023C34.7206 24.0051 35.2381 25.7499 35.2636 25.7755C35.3147 25.833 41.307 25.3601 41.3709 25.2961C41.3901 25.2706 40.7257 23.1806 39.8952 20.6368L38.3875 16.0222L38.707 15.741C39.7674 14.8142 40.6426 13.0246 40.8918 11.2542C41.0004 10.4808 40.9301 8.56339 40.764 7.85394C40.1252 5.07367 38.4323 2.91336 36.273 2.10805C35.3339 1.76291 35.4105 1.7693 30.9706 1.74373L26.8309 1.71817V13.3506ZM33.4174 7.22758C33.8709 7.45128 34.2415 8.0265 34.4395 8.81904C34.5737 9.34953 34.5545 11.0305 34.4076 11.5674C34.1329 12.5772 33.6473 13.1652 32.9191 13.3889L32.7721 13.4273V10.2315V7.02944L32.983 7.07418C33.0916 7.09336 33.2896 7.16366 33.4174 7.22758Z" fill="#666FE4" />
-                                    <path d="M41.0774 4.76085V7.73286H42.7064H44.3355V16.2974V24.8619H47.4019H50.4683V16.8087V8.75549H52.1293H53.7903V5.78348V2.81146L53.6498 2.7859C53.5156 2.75394 41.3329 1.78244 41.1604 1.78884C41.0966 1.78884 41.0774 2.42798 41.0774 4.76085Z" fill="#666FE4" />
-                                    <path d="M69.8881 14.0284V25.6288H73.0503H76.2126V14.0284V2.42797H73.0503H69.8881V14.0284Z" fill="#666FE4" />
-                                    <path d="M83.4328 3.23408C79.9703 3.41943 77.0891 3.57283 77.0316 3.57283C76.923 3.57922 76.9166 3.76457 76.9166 6.58319V9.58716H78.5457H80.1747V17.9919V26.3966H83.2731H86.3714V17.6404V8.8841H88.0963H89.8212V5.88013C89.8212 4.22475 89.802 2.88255 89.7765 2.88255C89.7445 2.88894 86.8953 3.04234 83.4328 3.23408Z" fill="#666FE4" />
-                                </svg>
-                            </div>
-                            <div className="flex-initial px-2 md:px-4">
-                                <svg width="20" height="13" viewBox="0 0 20 13" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <path d="M17.7111 0.5L19.9717 2.79889L10.4717 12.5L0.97168 2.79889L3.23225 0.5L10.4717 7.89269L17.7111 0.5Z" fill="#666FE4" />
-                                </svg>
+            {/* {mediaSorted.length === 0 && <Loader />}
+            mediaSorted.length >= 0 && () */}
+            <div className="flex flex-col h-full lg:flex-row">
+                <Sidebar />
+                {/* The homepage container */}
+                <div className="w-full pb-8 vod-content">
 
-                            </div>
-                            <div className="flex-1 pr-5 md:pr-10">
-                                <Search />
-                            </div>
-                            <div className="flex-initial ">
-                                <svg width="56" height="56" viewBox="0 0 56 56" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <rect x="56" y="56" width="56" height="56" rx="28" transform="rotate(-180 56 56)" fill="#F9F6FD" />
-                                    <circle cx="20" cy="28" r="2" fill="#666FE4" />
-                                    <circle cx="28" cy="28" r="2" fill="#666FE4" />
-                                    <circle cx="36" cy="28" r="2" fill="#666FE4" />
-                                </svg>
-                            </div>
-                            <div className="flex-initial px-5 ">
-                                <svg width="164" height="56" viewBox="0 0 164 56" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <rect width="164" height="56" rx="8" fill="#666FE4" />
-                                    <path d="M43.23 22.485C43.48 22.485 43.695 22.575 43.875 22.755C44.065 22.925 44.16 23.14 44.16 23.4V32.085C44.16 32.345 44.075 32.565 43.905 32.745C43.735 32.915 43.51 33 43.23 33C42.95 33 42.72 32.915 42.54 32.745C42.37 32.575 42.285 32.355 42.285 32.085V26.04L39.81 29.82C39.71 29.94 39.595 30.04 39.465 30.12C39.335 30.19 39.2 30.22 39.06 30.21C38.93 30.22 38.8 30.19 38.67 30.12C38.54 30.04 38.425 29.94 38.325 29.82L35.85 25.905V32.085C35.85 32.355 35.77 32.575 35.61 32.745C35.45 32.915 35.24 33 34.98 33C34.73 33 34.525 32.915 34.365 32.745C34.205 32.565 34.125 32.345 34.125 32.085V23.4C34.125 23.15 34.215 22.935 34.395 22.755C34.585 22.575 34.81 22.485 35.07 22.485C35.22 22.485 35.37 22.525 35.52 22.605C35.67 22.675 35.785 22.77 35.865 22.89L39.135 27.915L42.495 22.89C42.575 22.77 42.68 22.675 42.81 22.605C42.95 22.525 43.09 22.485 43.23 22.485ZM52.22 24.975C52.48 24.975 52.695 25.065 52.865 25.245C53.035 25.415 53.12 25.63 53.12 25.89V32.265C53.12 33.565 52.77 34.51 52.07 35.1C51.37 35.7 50.435 36 49.265 36C48.925 36 48.575 35.97 48.215 35.91C47.865 35.86 47.58 35.79 47.36 35.7C46.87 35.49 46.625 35.2 46.625 34.83C46.625 34.76 46.64 34.68 46.67 34.59C46.73 34.37 46.83 34.2 46.97 34.08C47.12 33.97 47.28 33.915 47.45 33.915C47.55 33.915 47.645 33.93 47.735 33.96C47.805 33.99 47.925 34.04 48.095 34.11C48.275 34.18 48.46 34.235 48.65 34.275C48.85 34.325 49.055 34.35 49.265 34.35C49.975 34.35 50.495 34.195 50.825 33.885C51.165 33.585 51.335 33.065 51.335 32.325V32.175C50.815 32.825 50.005 33.15 48.905 33.15C48.385 33.15 47.925 33.03 47.525 32.79C47.125 32.54 46.815 32.2 46.595 31.77C46.375 31.33 46.265 30.83 46.265 30.27V25.89C46.265 25.63 46.35 25.415 46.52 25.245C46.69 25.065 46.905 24.975 47.165 24.975C47.425 24.975 47.64 25.065 47.81 25.245C47.98 25.415 48.065 25.63 48.065 25.89V29.73C48.065 30.36 48.2 30.815 48.47 31.095C48.75 31.365 49.155 31.5 49.685 31.5C50.205 31.5 50.605 31.35 50.885 31.05C51.175 30.74 51.32 30.3 51.32 29.73V25.89C51.32 25.63 51.405 25.415 51.575 25.245C51.745 25.065 51.96 24.975 52.22 24.975ZM67.0279 22.5C67.3079 22.5 67.5329 22.585 67.7029 22.755C67.8729 22.925 67.9579 23.145 67.9579 23.415V32.085C67.9579 32.345 67.8679 32.565 67.6879 32.745C67.5079 32.915 67.2779 33 66.9979 33C66.7479 33 66.5329 32.915 66.3529 32.745C66.1729 32.565 66.0829 32.345 66.0829 32.085V28.59H61.2979V32.085C61.2979 32.345 61.2029 32.565 61.0129 32.745C60.8229 32.915 60.6029 33 60.3529 33C60.0829 33 59.8579 32.915 59.6779 32.745C59.5079 32.565 59.4229 32.345 59.4229 32.085V23.415C59.4229 23.155 59.5129 22.94 59.6929 22.77C59.8729 22.59 60.1029 22.5 60.3829 22.5C60.6329 22.5 60.8479 22.59 61.0279 22.77C61.2079 22.94 61.2979 23.155 61.2979 23.415V26.865H66.0829V23.415C66.0829 23.155 66.1779 22.94 66.3679 22.77C66.5579 22.59 66.7779 22.5 67.0279 22.5ZM71.7491 32.085C71.7491 32.345 71.6641 32.565 71.4941 32.745C71.3241 32.915 71.1091 33 70.8491 33C70.5891 33 70.3741 32.915 70.2041 32.745C70.0341 32.565 69.9491 32.345 69.9491 32.085V25.89C69.9491 25.63 70.0341 25.415 70.2041 25.245C70.3741 25.065 70.5891 24.975 70.8491 24.975C71.1091 24.975 71.3241 25.065 71.4941 25.245C71.6641 25.415 71.7491 25.63 71.7491 25.89V32.085ZM70.8341 24C70.4941 24 70.2541 23.945 70.1141 23.835C69.9741 23.725 69.9041 23.53 69.9041 23.25V22.965C69.9041 22.685 69.9791 22.49 70.1291 22.38C70.2791 22.27 70.5191 22.215 70.8491 22.215C71.1991 22.215 71.4441 22.27 71.5841 22.38C71.7241 22.49 71.7941 22.685 71.7941 22.965V23.25C71.7941 23.54 71.7191 23.74 71.5691 23.85C71.4291 23.95 71.1841 24 70.8341 24ZM80.244 24.825C80.504 24.825 80.719 24.91 80.889 25.08C81.059 25.25 81.144 25.47 81.144 25.74V32.415C81.144 33.635 80.764 34.535 80.004 35.115C79.254 35.705 78.299 36 77.139 36C76.839 36 76.424 35.95 75.894 35.85C75.374 35.75 75.004 35.65 74.784 35.55C74.294 35.34 74.049 35.05 74.049 34.68C74.049 34.61 74.064 34.53 74.094 34.44C74.154 34.22 74.254 34.05 74.394 33.93C74.544 33.82 74.704 33.765 74.874 33.765C74.974 33.765 75.069 33.78 75.159 33.81C75.199 33.82 75.444 33.91 75.894 34.08C76.354 34.26 76.769 34.35 77.139 34.35C77.879 34.35 78.429 34.2 78.789 33.9C79.159 33.61 79.344 33.195 79.344 32.655V31.935C79.124 32.245 78.789 32.525 78.339 32.775C77.899 33.025 77.424 33.15 76.914 33.15C76.214 33.15 75.574 32.975 74.994 32.625C74.424 32.265 73.974 31.77 73.644 31.14C73.314 30.51 73.149 29.795 73.149 28.995C73.149 28.195 73.309 27.48 73.629 26.85C73.959 26.21 74.404 25.715 74.964 25.365C75.534 25.005 76.164 24.825 76.854 24.825C77.364 24.825 77.839 24.93 78.279 25.14C78.719 25.35 79.074 25.605 79.344 25.905V25.74C79.344 25.48 79.429 25.265 79.599 25.095C79.769 24.915 79.984 24.825 80.244 24.825ZM77.139 31.5C77.809 31.5 78.354 31.265 78.774 30.795C79.204 30.315 79.419 29.715 79.419 28.995C79.419 28.275 79.204 27.675 78.774 27.195C78.354 26.715 77.809 26.475 77.139 26.475C76.469 26.475 75.924 26.715 75.504 27.195C75.084 27.675 74.874 28.275 74.874 28.995C74.874 29.715 75.084 30.315 75.504 30.795C75.924 31.265 76.469 31.5 77.139 31.5ZM87.3062 24.825C88.2962 24.825 88.9812 25.13 89.3612 25.74C89.7412 26.34 89.9312 27.175 89.9312 28.245V32.085C89.9312 32.345 89.8462 32.565 89.6762 32.745C89.5062 32.915 89.2912 33 89.0312 33C88.7712 33 88.5562 32.915 88.3862 32.745C88.2162 32.565 88.1312 32.345 88.1312 32.085V28.245C88.1312 27.695 88.0212 27.265 87.8012 26.955C87.5912 26.635 87.2162 26.475 86.6762 26.475C86.1162 26.475 85.6762 26.645 85.3562 26.985C85.0462 27.315 84.8912 27.735 84.8912 28.245V32.085C84.8912 32.345 84.8062 32.565 84.6362 32.745C84.4662 32.915 84.2512 33 83.9912 33C83.7312 33 83.5162 32.915 83.3462 32.745C83.1762 32.565 83.0912 32.345 83.0912 32.085V22.815C83.0912 22.555 83.1762 22.34 83.3462 22.17C83.5162 21.99 83.7312 21.9 83.9912 21.9C84.2512 21.9 84.4662 21.99 84.6362 22.17C84.8062 22.34 84.8912 22.555 84.8912 22.815V26.13C85.1612 25.77 85.5012 25.465 85.9112 25.215C86.3312 24.955 86.7962 24.825 87.3062 24.825ZM93.8303 32.085C93.8303 32.345 93.7403 32.565 93.5603 32.745C93.3903 32.915 93.1753 33 92.9153 33C92.6653 33 92.4553 32.915 92.2853 32.745C92.1153 32.565 92.0303 32.345 92.0303 32.085V22.815C92.0303 22.555 92.1153 22.34 92.2853 22.17C92.4653 21.99 92.6853 21.9 92.9453 21.9C93.1953 21.9 93.4053 21.99 93.5753 22.17C93.7453 22.34 93.8303 22.555 93.8303 22.815V32.085ZM97.6182 32.085C97.6182 32.345 97.5332 32.565 97.3632 32.745C97.1932 32.915 96.9782 33 96.7182 33C96.4582 33 96.2432 32.915 96.0732 32.745C95.9032 32.565 95.8182 32.345 95.8182 32.085V25.89C95.8182 25.63 95.9032 25.415 96.0732 25.245C96.2432 25.065 96.4582 24.975 96.7182 24.975C96.9782 24.975 97.1932 25.065 97.3632 25.245C97.5332 25.415 97.6182 25.63 97.6182 25.89V32.085ZM96.7032 24C96.3632 24 96.1232 23.945 95.9832 23.835C95.8432 23.725 95.7732 23.53 95.7732 23.25V22.965C95.7732 22.685 95.8482 22.49 95.9982 22.38C96.1482 22.27 96.3882 22.215 96.7182 22.215C97.0682 22.215 97.3132 22.27 97.4532 22.38C97.5932 22.49 97.6632 22.685 97.6632 22.965V23.25C97.6632 23.54 97.5882 23.74 97.4382 23.85C97.2982 23.95 97.0532 24 96.7032 24ZM106.113 24.825C106.373 24.825 106.588 24.91 106.758 25.08C106.928 25.25 107.013 25.47 107.013 25.74V32.415C107.013 33.635 106.633 34.535 105.873 35.115C105.123 35.705 104.168 36 103.008 36C102.708 36 102.293 35.95 101.763 35.85C101.243 35.75 100.873 35.65 100.653 35.55C100.163 35.34 99.9182 35.05 99.9182 34.68C99.9182 34.61 99.9332 34.53 99.9632 34.44C100.023 34.22 100.123 34.05 100.263 33.93C100.413 33.82 100.573 33.765 100.743 33.765C100.843 33.765 100.938 33.78 101.028 33.81C101.068 33.82 101.313 33.91 101.763 34.08C102.223 34.26 102.638 34.35 103.008 34.35C103.748 34.35 104.298 34.2 104.658 33.9C105.028 33.61 105.213 33.195 105.213 32.655V31.935C104.993 32.245 104.658 32.525 104.208 32.775C103.768 33.025 103.293 33.15 102.783 33.15C102.083 33.15 101.443 32.975 100.863 32.625C100.293 32.265 99.8432 31.77 99.5132 31.14C99.1832 30.51 99.0182 29.795 99.0182 28.995C99.0182 28.195 99.1782 27.48 99.4982 26.85C99.8282 26.21 100.273 25.715 100.833 25.365C101.403 25.005 102.033 24.825 102.723 24.825C103.233 24.825 103.708 24.93 104.148 25.14C104.588 25.35 104.943 25.605 105.213 25.905V25.74C105.213 25.48 105.298 25.265 105.468 25.095C105.638 24.915 105.853 24.825 106.113 24.825ZM103.008 31.5C103.678 31.5 104.223 31.265 104.643 30.795C105.073 30.315 105.288 29.715 105.288 28.995C105.288 28.275 105.073 27.675 104.643 27.195C104.223 26.715 103.678 26.475 103.008 26.475C102.338 26.475 101.793 26.715 101.373 27.195C100.953 27.675 100.743 28.275 100.743 28.995C100.743 29.715 100.953 30.315 101.373 30.795C101.793 31.265 102.338 31.5 103.008 31.5ZM113.175 24.825C114.165 24.825 114.85 25.13 115.23 25.74C115.61 26.34 115.8 27.175 115.8 28.245V32.085C115.8 32.345 115.715 32.565 115.545 32.745C115.375 32.915 115.16 33 114.9 33C114.64 33 114.425 32.915 114.255 32.745C114.085 32.565 114 32.345 114 32.085V28.245C114 27.695 113.89 27.265 113.67 26.955C113.46 26.635 113.085 26.475 112.545 26.475C111.985 26.475 111.545 26.645 111.225 26.985C110.915 27.315 110.76 27.735 110.76 28.245V32.085C110.76 32.345 110.675 32.565 110.505 32.745C110.335 32.915 110.12 33 109.86 33C109.6 33 109.385 32.915 109.215 32.745C109.045 32.565 108.96 32.345 108.96 32.085V22.815C108.96 22.555 109.045 22.34 109.215 22.17C109.385 21.99 109.6 21.9 109.86 21.9C110.12 21.9 110.335 21.99 110.505 22.17C110.675 22.34 110.76 22.555 110.76 22.815V26.13C111.03 25.77 111.37 25.465 111.78 25.215C112.2 24.955 112.665 24.825 113.175 24.825ZM121.619 31.305C121.759 31.305 121.884 31.375 121.994 31.515C122.114 31.645 122.174 31.82 122.174 32.04C122.174 32.31 122.024 32.54 121.724 32.73C121.434 32.91 121.104 33 120.734 33C120.114 33 119.589 32.87 119.159 32.61C118.739 32.34 118.529 31.775 118.529 30.915V26.775H117.839C117.599 26.775 117.399 26.695 117.239 26.535C117.079 26.375 116.999 26.175 116.999 25.935C116.999 25.705 117.079 25.515 117.239 25.365C117.399 25.205 117.599 25.125 117.839 25.125H118.529V24.165C118.529 23.905 118.614 23.69 118.784 23.52C118.964 23.34 119.184 23.25 119.444 23.25C119.694 23.25 119.904 23.34 120.074 23.52C120.244 23.69 120.329 23.905 120.329 24.165V25.125H121.394C121.634 25.125 121.834 25.205 121.994 25.365C122.154 25.525 122.234 25.725 122.234 25.965C122.234 26.195 122.154 26.39 121.994 26.55C121.834 26.7 121.634 26.775 121.394 26.775H120.329V30.84C120.329 31.05 120.384 31.205 120.494 31.305C120.604 31.395 120.754 31.44 120.944 31.44C121.024 31.44 121.134 31.42 121.274 31.38C121.394 31.33 121.509 31.305 121.619 31.305ZM123.715 31.755C123.625 31.635 123.58 31.485 123.58 31.305C123.58 31.045 123.71 30.825 123.97 30.645C124.11 30.555 124.255 30.51 124.405 30.51C124.615 30.51 124.815 30.605 125.005 30.795C125.305 31.105 125.615 31.335 125.935 31.485C126.265 31.635 126.66 31.71 127.12 31.71C127.9 31.68 128.29 31.385 128.29 30.825C128.29 30.535 128.155 30.315 127.885 30.165C127.625 30.015 127.215 29.865 126.655 29.715C126.075 29.555 125.6 29.395 125.23 29.235C124.87 29.075 124.555 28.835 124.285 28.515C124.025 28.185 123.895 27.745 123.895 27.195C123.895 26.715 124.03 26.3 124.3 25.95C124.58 25.59 124.94 25.315 125.38 25.125C125.83 24.925 126.305 24.825 126.805 24.825C127.315 24.825 127.825 24.92 128.335 25.11C128.845 25.3 129.245 25.58 129.535 25.95C129.665 26.11 129.73 26.29 129.73 26.49C129.73 26.72 129.635 26.91 129.445 27.06C129.305 27.17 129.145 27.225 128.965 27.225C128.745 27.225 128.565 27.155 128.425 27.015C128.245 26.795 128.01 26.62 127.72 26.49C127.44 26.36 127.12 26.295 126.76 26.295C126 26.295 125.62 26.57 125.62 27.12C125.62 27.34 125.69 27.52 125.83 27.66C125.97 27.79 126.145 27.895 126.355 27.975C126.575 28.045 126.88 28.135 127.27 28.245C127.81 28.395 128.26 28.55 128.62 28.71C128.98 28.87 129.285 29.115 129.535 29.445C129.785 29.775 129.91 30.21 129.91 30.75C129.91 31.24 129.765 31.67 129.475 32.04C129.185 32.4 128.81 32.675 128.35 32.865C127.9 33.055 127.435 33.15 126.955 33.15C126.295 33.15 125.68 33.04 125.11 32.82C124.54 32.6 124.075 32.245 123.715 31.755Z" fill="white" />
-                                </svg>
-                            </div>
-                        </div>
-                        <div className='w-full m-auto overflow-hidden shadow'>
-                            <Swiper {...params} containerClass='m-auto max-w-3xl flex flex-row px-12 py-4 overflow-hidden'>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>All</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Solo</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Random</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Quadra</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Quadra</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Quadra</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Quadra</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Quadra</div>
-                                <div className='p-2 text-center bg-gray-200 rounded-full shadow'>Snipe</div>
-                            </Swiper>
-                        </div>
-
-                        <div className="flex my-8 ml-8 text-2xl font-bold text-primary-500">Top Highlights</div>
-                        {/* <VodTabs> */}
-                        <div className="flex my-4 text-sm">
-                            <div className={currentSince.current === 1 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500 w" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(1)}>Last 2 Days</div>
-                            <div className={currentSince.current === 7 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(7)}>Last Week</div>
-                            <div className={currentSince.current === 14 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(14)}>Last 2 Weeks</div>
-                        </div>
-                        <TopRated
-                            mediaSorted={mediaSorted}
-                            mediaById={mediaById}
-                            gaEvent="Home Page::Top rated"
-                        />
-                        {/* </VodTabs> */}
-                        <div className="flex flex-col flex-wrap my-4 sm:flex-row">
-
-                            <div className="p-2 sm:w-1/2 xl:w-1/3">
-                                <div>
-                                    <a href="#">
-                                        <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
-                                            <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
-                                            <div className="flex items-center">
-                                                <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
-                                                <div className="">
-                                                    <div className="font-bold text-purple-500">Shadowburn</div>
-
-                                                    <div className="mt-1 text-sm font-medium">2k viewers and ur pwning</div>
-                                                </div>
-                                            </div>
-
-                                        </div>
-                                    </a>
-                                </div>
-                            </div>
-                            <div className="p-2 sm:w-1/2 xl:w-1/3">
-                                <div>
-                                    <a href="#">
-                                        <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
-                                            <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
-                                            <div className="flex items-center">
-                                                <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
-                                                <div className="">
-                                                    <div className="font-bold text-purple-500">Shadowburn</div>
-
-                                                    <div className="mt-1 text-sm font-medium">How are you u winning</div>
-                                                </div>
-                                            </div>
-
-                                        </div>
-                                    </a>
-                                </div>
-                            </div>
-                            <div className="p-2 sm:w-1/2 xl:w-1/3">
-                                <div>
-                                    <a href="#">
-                                        <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
-                                            <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
-                                            <div className="flex items-center">
-                                                <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
-                                                <div className="">
-                                                    <div className="font-bold text-purple-500">Shadowburn</div>
-
-                                                    <div className="mt-1 text-sm font-medium">How are you u winning</div>
-                                                </div>
-                                            </div>
-
-                                        </div>
-                                    </a>
-                                </div>
-                            </div>
-                            <div className="p-2 sm:w-1/2 xl:w-1/3">
-                                <div>
-                                    <a href="#">
-                                        <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
-                                            <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
-                                            <div className="flex items-center">
-                                                <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
-                                                <div className="">
-                                                    <div className="font-bold text-purple-500">Shadowburn</div>
-
-                                                    <div className="mt-1 text-sm font-medium">How are you u winning</div>
-                                                </div>
-                                            </div>
-
-                                        </div>
-                                    </a>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div className="flex flex-col items-center justify-center w-full mt-12 cursor-pointer">
-                            <div className="mb-2 text-sm font-bold">
-                                Load more
-                </div>
-                            <div className="flex items-center justify-center w-16 h-16 transition duration-300 ease-in-out rounded-full bg-primary-800 hover:bg-indigo-800">
-                                <svg className="w-6 h-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 21 24">
-                                    <defs />
-                                    <path fill="white" fillRule="evenodd" d="M19.6347 12.2744c.875.8721.8762 2.2872.0027 3.1608l-7.9088 7.9088c-.8736.8736-2.29119.8748-3.16626.0027-.02961-.0295-.05823-.0596-.08584-.0903L.655868 15.4357c-.873579-.8736-.874792-2.2911-.002705-3.1662.872087-.8751 2.287237-.8763 3.160817-.0027l4.09257 4.0926V2.24077c0-1.23543 1.00151-2.2386572 2.23695-2.24076711C11.3789-.00210702 12.3804.997697 12.3804 2.23313V16.3651l4.088-4.088c.8736-.8736 2.2912-.8748 3.1663-.0027z" clipRule="evenodd" />
-                                </svg>
-
-                            </div>
-                        </div>
-
+                    <Header />
+                    <Tags />
+                    <div className="flex my-8 ml-8 text-2xl font-bold text-primary-500">Top Highlights</div>
+                    {/* <VodTabs> */}
+                    <div className="flex my-4 text-sm">
+                        <div className={currentSince.current === 1 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500 w" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(1)}>Last 2 Days</div>
+                        <div className={currentSince.current === 7 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(7)}>Last Week</div>
+                        <div className={currentSince.current === 14 ? "ml-8 pb-2 font-bold border-b-4 border-solid cursor-pointer text-primary-500 border-primary-500" : 'ml-8 text-gray-600 cursor-pointer'} onClick={() => sortBySince(14)}>Last 2 Weeks</div>
                     </div>
+                    <TopRated
+                        mediaSorted={mediaSorted}
+                        mediaById={mediaById}
+                        gaEvent="Home Page::Top rated"
+                    />
+                    {/* </VodTabs> */}
+                    <div className="flex flex-col flex-wrap my-4 sm:flex-row">
+                        <div className="p-2 sm:w-1/2 xl:w-1/3">
+                            <div>
+                                <a href="#">
+                                    <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
+                                        <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
+                                        <div className="flex items-center">
+                                            <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
+                                            <div className="">
+                                                <div className="font-bold text-purple-500">Shadowburn</div>
+
+                                                <div className="mt-1 text-sm font-medium">2k viewers and ur pwning</div>
+                                            </div>
+                                        </div>
+
+                                    </div>
+                                </a>
+                            </div>
+                        </div>
+                        <div className="p-2 sm:w-1/2 xl:w-1/3">
+                            <div>
+                                <a href="#">
+                                    <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
+                                        <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
+                                        <div className="flex items-center">
+                                            <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
+                                            <div className="">
+                                                <div className="font-bold text-purple-500">Shadowburn</div>
+
+                                                <div className="mt-1 text-sm font-medium">How are you u winning</div>
+                                            </div>
+                                        </div>
+
+                                    </div>
+                                </a>
+                            </div>
+                        </div>
+                        <div className="p-2 sm:w-1/2 xl:w-1/3">
+                            <div>
+                                <a href="#">
+                                    <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
+                                        <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
+                                        <div className="flex items-center">
+                                            <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
+                                            <div className="">
+                                                <div className="font-bold text-purple-500">Shadowburn</div>
+
+                                                <div className="mt-1 text-sm font-medium">How are you u winning</div>
+                                            </div>
+                                        </div>
+
+                                    </div>
+                                </a>
+                            </div>
+                        </div>
+                        <div className="p-2 sm:w-1/2 xl:w-1/3">
+                            <div>
+                                <a href="#">
+                                    <div className="relative p-2 text-gray-800 bg-white rounded-lg shadow hover:shadow-lg">
+                                        <img src="https://i.ytimg.com/vi/jtFRndWvu5A/maxresdefault.jpg" className="object-cover w-full h-32 mb-2 rounded-lg rounded-b-none" />
+                                        <div className="flex items-center">
+                                            <img src="https://media.forgecdn.net/avatars/100/287/636314173868302488.jpeg" className="object-cover object-center w-12 h-12 mr-2 border-4 border-white rounded-full" />
+                                            <div className="">
+                                                <div className="font-bold text-purple-500">Shadowburn</div>
+
+                                                <div className="mt-1 text-sm font-medium">How are you u winning</div>
+                                            </div>
+                                        </div>
+
+                                    </div>
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+
+                    <div className="flex flex-col items-center justify-center w-full mt-12 cursor-pointer">
+                        <div className="mb-2 text-sm font-bold">
+                            Load more
                 </div>
-            )}
+                        <div className="flex items-center justify-center w-16 h-16 transition duration-300 ease-in-out rounded-full bg-primary-800 hover:bg-indigo-800">
+                            <svg className="w-6 h-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 21 24">
+                                <defs />
+                                <path fill="white" fillRule="evenodd" d="M19.6347 12.2744c.875.8721.8762 2.2872.0027 3.1608l-7.9088 7.9088c-.8736.8736-2.29119.8748-3.16626.0027-.02961-.0295-.05823-.0596-.08584-.0903L.655868 15.4357c-.873579-.8736-.874792-2.2911-.002705-3.1662.872087-.8751 2.287237-.8763 3.160817-.0027l4.09257 4.0926V2.24077c0-1.23543 1.00151-2.2386572 2.23695-2.24076711C11.3789-.00210702 12.3804.997697 12.3804 2.23313V16.3651l4.088-4.088c.8736-.8736 2.2912-.8748 3.1663-.0027z" clipRule="evenodd" />
+                            </svg>
+
+                        </div>
+                    </div>
+
+                </div>
+            </div>
+
         </>
     );
 }
